+++ src/main/java/org/usfirst/frc/team319/robot/OI.java
@@ -13,8 +13,8 @@ import org.usfirst.frc.team319.robot.commands.bba.BbaConfigSoftLimits;
 import org.usfirst.frc.team319.robot.commands.bba.ResetBbaPosition;
 import org.usfirst.frc.team319.robot.commands.carriage.CollectorSetSpeed;
 import org.usfirst.frc.team319.robot.commands.carriage.PassthroughSpit;
-import org.usfirst.frc.team319.robot.commands.hatchCollector.HatchCollectorCollect;
-import org.usfirst.frc.team319.robot.commands.hatchCollector.HatchCollectorPlace;
+import org.usfirst.frc.team319.robot.commands.hatch_collector.HatchCollectorCollect;
+import org.usfirst.frc.team319.robot.commands.hatch_collector.HatchCollectorPlace;
 import org.usfirst.frc.team319.robot.commands.robot.CollectCargo;
 import org.usfirst.frc.team319.robot.commands.robot.GoToCargoShipPose;
 import org.usfirst.frc.team319.robot.commands.robot.GoToCollectPose;
@@ -25,8 +25,6 @@ import org.usfirst.frc.team319.robot.commands.robot.GoToSafePose;
 import org.usfirst.frc.team319.robot.commands.robot.SetDriveMode;
 import org.usfirst.frc.team319.robot.commands.robot.SpitCargo;
 import org.usfirst.frc.team319.robot.commands.robot.StartClimbLevelTwoLowLockMode;
-import org.usfirst.frc.team319.robot.commands.robot.StartClimbLevelTwoMode;
-import org.usfirst.frc.team319.robot.commands.robot.StartClimbMode;
 import org.usfirst.frc.team319.robot.commands.robot.StartLowLockClimbMode;
 import org.usfirst.frc.team319.robot.commands.robot.StopCollectCargo;
