+++ src/main/java/org/usfirst/frc/team319/robot/commands/hatch_collector/HatchCollectorOpen.java
@@ -5,14 +5,14 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package org.usfirst.frc.team319.robot.commands.hatchCollector;
+package org.usfirst.frc.team319.robot.commands.hatch_collector;
 
 import org.usfirst.frc.team319.robot.Robot;
 
 import edu.wpi.first.wpilibj.command.Command;
 
-public class HatchCollectorToggle extends Command {
-  public HatchCollectorToggle() {
+public class HatchCollectorOpen extends Command {
+  public HatchCollectorOpen() {
     // Use requires() here to declare subsystem dependencies
     requires(Robot.pneumatics);
   }
@@ -25,11 +25,7 @@ public class HatchCollectorToggle extends Command {
   // Called repeatedly when this Command is scheduled to run
   @Override
   protected void execute() {
-    if (Robot.carriage.isFingerOpen()) {
-      Robot.pneumatics.fingerClose();
-    } else {
-      Robot.pneumatics.fingerOpen();
-    }
+    Robot.pneumatics.fingerOpen();
   }
 
   // Make this return true when this Command no longer needs to run execute()
